Refactor Captain of Bass application structure and update dependencies
- Downgraded "tw-animate-css" version in bun.lock for compatibility.
- Changed TypeScript module resolution from "node16" to "node" for broader support.
- Removed unused Tailwind CSS configuration files from multiple applications.
- Updated global CSS to include necessary imports for Tailwind and animation utilities.
- Refactored component imports in layout and music pages for improved organization.
- Deleted redundant footer and header components to streamline the codebase.
