+++ apps/captainofbass.com/tailwind.config.ts
@@ -1,7 +1,6 @@
-import type { Config } from 'tailwindcss'
 import { tailwindConfig } from '@mikeallisonjs/shared-react-lib'
 
-const config: Config = {
+const config = {
   ...tailwindConfig
 }
