+++ packages/shared-react-lib/tsconfig.json
@@ -3,6 +3,7 @@
     "jsx": "react-jsx",
     "allowJs": false,
     "esModuleInterop": false,
+    "moduleResolution": "bundler",
     "allowSyntheticDefaultImports": true,
     "strict": true
   },
